Move IRIS import tests (TestIrisImports) to plugin suite
